updates from pr comments
